Fix #1096
Fixes an issue with portable serialization resulting in different byte arrays. The cause
was a bytearray which wasn't cleaned correctly.
