+++ hazelcast/src/main/java/com/hazelcast/nio/BufferObjectDataOutput.java
@@ -34,6 +34,8 @@ public interface BufferObjectDataOutput extends ObjectDataOutput, Closeable {
 
     void writeByte(int position, final int v) throws IOException;
 
+    void writeZeroBytes(int count);
+
     void writeChar(int position, final int v) throws IOException;
 
     void writeDouble(int position, final double v) throws IOException;
